chain ids are now uint64 as per https://eips.ethereum.org/EIPS/eip-2294

+++ packages/chain/contracts/token/ITokenProxy.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.20;
 
 interface ITokenProxy {
-    function chain() external returns (uint);
+    function chain() external returns (uint64);
     function token() external returns (address);
 }
